improved and simplified configuration files
