+++ tasks/options/webpack-dev-server.js
@@ -1,4 +1,13 @@
 module.exports = function(grunt) {
+    var isModule = require('fs').existsSync('./node_modules/superdesk-core');
+    var rewrite = {
+        target: 'http://localhost:9000',
+        rewrite: function(req) {
+            'use strict';
+            req.url = 'node_modules/superdesk-core' + req.url;
+        }
+    };
+
     return {
         options: {
             webpack: require('../../webpack.config.js')(grunt, true),
@@ -10,6 +19,10 @@ module.exports = function(grunt) {
         },
         start: {
             keepAlive: true,
+            proxy: isModule ? {
+                '/scripts/*': rewrite,
+                '/images/*': rewrite
+            } : {},
             webpack: {
                 devtool: 'eval',
                 debug: true
